Add metrics for SLI monitoring
